Convert Ruby tests to smoke tests
