+++ .dockerignore
@@ -1,4 +1,4 @@
-test/
+tests/
 scripts/
 node_modules/
 .appends
