+++ .gitignore
@@ -2,3 +2,7 @@
 /tmp/
 
 node_modules
+
+/tests/**/representation.txt
+/tests/**/representation.json
+/tests/**/mapping.json
