+++ tests/examples/basic/basic.rb
@@ -1,0 +1,5 @@
+class TwoFer
+  def two_fer
+    "foo"
+  end
+end
