+++ tests/mapping/arguments/arguments.rb
@@ -1,0 +1,2 @@
+def somefunc(foo, bar)
+end
