+++ tests/mapping/attr_accessors/attr_accessors.rb
@@ -1,0 +1,6 @@
+class Foobar
+  attr_reader :foobar
+  attr_reader "other1", "other2"
+  attr_writer :barfood
+  attr_accessor :misc
+end
