+++ tests/mapping/class/class.rb
@@ -1,0 +1,2 @@
+class Foobar
+end
