+++ tests/mapping/class_variables/class_variables.rb
@@ -1,0 +1,3 @@
+class Foobar
+  foobar = true
+end
