+++ tests/mapping/local_variables/local_variables.rb
@@ -1,0 +1,2 @@
+foobar = true
+puts barfoo
