+++ tests/mapping/method/method.rb
@@ -1,0 +1,3 @@
+class Foobar
+end
+BarFoo
