+++ tests/mapping/module/module.rb
@@ -1,0 +1,3 @@
+module Foobar
+end
+BarFoo
