+++ tests/naming_normalisation/args/args.rb
@@ -1,0 +1,5 @@
+foo = 1
+bar = 2
+def somefunc(param1, bar)
+  return param1 + bar
+end
