+++ tests/naming_normalisation/class_name/class_name.rb
@@ -1,0 +1,1 @@
+class Foobar; end
