+++ tests/naming_normalisation/complex/complex.rb
@@ -1,0 +1,10 @@
+class TwoFer
+  def two_fer
+    "foo"
+  end
+
+  def foobar
+    two_fer = "cat"
+    return TwoFer.two_fer
+  end
+end
