+++ tests/naming_normalisation/complex_method/complex_method.rb
@@ -1,0 +1,5 @@
+def initialize(a, b)
+  @alphabet = ('a'..'z').to_a
+  @key = [a, b]; coprime_check(key[0])
+  @cipherbet = make_cipherbet.join
+end
