+++ tests/naming_normalisation/global_variable/global_variable.rb
@@ -1,0 +1,1 @@
+$foo = bar
