+++ tests/naming_normalisation/if/if.rb
@@ -1,0 +1,1 @@
+foo = true; if foo then 'bar' else nil end
