+++ tests/naming_normalisation/instance_variable/instance_variable.rb
@@ -1,0 +1,1 @@
+@foo = bar
