+++ tests/naming_normalisation/method_call/method_call.rb
@@ -1,0 +1,2 @@
+def bar; end
+foo.bar
