+++ tests/naming_normalisation/method_name/method_name.rb
@@ -1,0 +1,1 @@
+def foobar; end
