+++ tests/naming_normalisation/non_defined_method_call/non_defined_method_call.rb
@@ -1,0 +1,1 @@
+foo = []; foo.each
