+++ tests/naming_normalisation/return/return.rb
@@ -1,0 +1,1 @@
+foo = true; return foo
