+++ tests/naming_normalisation/variable_name/variable_name.rb
@@ -1,0 +1,1 @@
+foobar = true
